Fix tests that fail due to a MethodError
In 0.7, a byte array literal is of type CodeUnits, which means that
methods expecting a Vector{UInt8} fail when they get passed a string
constructed with the b_str macro.

While perhaps the API should be updated, for now this workaround where
we convert all the strings explicitly to Vectors in the tests is
sufficicent to get the tests to pass.